Get the code building on Windows again.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@1366561 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/decaf/util/BitSet.cpp
@@ -541,7 +541,7 @@ int BitSet::length() const {
         return 0;
     }
     int i = ELM_SIZE - 1;
-    long val = bits[idx];
+    unsigned long long val = bits[idx];
     while ((val & (TWO_N_ARRAY[i])) == 0 && i > 0) {
         i--;
     }
